Make a working Etch-A-Sketch using HTML, CSS and JS
Create a webpage with a 16x16 grid of square divs. Create the divs using JavaScript. Don’t try making them by hand with copy and pasting in your HTML file! Set up a hover effect so that the grid divs change color when your mouse passes over them, leaving a (pixelated) trail through your grid like a pen would.

Add a button to create a new grid with an input for the number of rows and colums in the grid. The existing grid should be removed and a new grid should be generated in the same total space as before so that you’ve got a new sketch pad. Set the limit for the user input to a maximum of 100. A larger number of squares results in more computer resources being used, potentially causing delays, freezing, or crashing that we want to prevent.

Add buttons to clear the color in the grid, let the color that is colored be random and to erase parts of the grid.

My excuses for this big commit, as I forgot to commit earlier!